Completed up to CRUD operation and search functionality and started implementing authentication.
Completed Book listing functionality that  shows all the list of book.
Completed pagination functionality that display the limited number of books on one page.
Completed search functionality that search for books based on Title, Author, or Genre.
Started Implementing authentication in that implemented Account controller and Login and Sign in page.